+++ kfinance/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v7.0.0
+- Include `status`, `start_date`, `target`, and `buyers` fields in `get_mergers_from_identifiers` 
+  tool call response and `MergerOrAcquisition` object.
+- Replace merger info tool call with a batch version named `get_mergers_info_from_transaction_ids`,
+  which also supports the option of including advisors in the response
+- Breaking changes:
+  - `merger_title` field has been removed from the `MergerOrAcquisition` object
+  type and the `MergerSummary` type, since the newly added fields make that field
+  redundant.
+  - `get_merger_info_from_transaction_id` tool call has been removed.
+
 ## v6.5.1
 - Change M&A tool descriptions and prompts.py for M&A performance improvement